Improve comm data link.
* Removed comm send timeout.  Implement in app and higher levels as needed.
* Added process_request callback for more reliable OS integration.
* Improved multi-threaded performance and fixed potential deadlock.
  Clarified that fbp_dl_send is the only thread-safe function.
* Added fbp_dl_process and removed fbp_evm integration to give better
  decoupling and performance.
* Added explicit buffer size to framer construct_data.
* Fixed stream_tester.

Also added fbp_rbu64_is_empty

+++ src/topic.c
@@ -32,7 +32,7 @@ void fbp_topic_append(struct fbp_topic_s * topic, const char * subtopic) {
     }
     FBP_ASSERT(t < topic_end);
     *t = 0;
-    topic->length = t - topic->topic;
+    topic->length = (uint8_t) (t - topic->topic);
 }
 
 void fbp_topic_set(struct fbp_topic_s * topic, const char * str) {
